requirements.txt
# ── Core Framework ────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0
pydantic>=2.0
email-validator>=2.0.0
python-dotenv>=1.0.0

# ── AI / LLM ─────────────────────────────────────────────
langgraph==0.4.7
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# ── Database ──────────────────────────────────────────────
sqlalchemy>=2.0.0

# ── Auth ──────────────────────────────────────────────────
bcrypt>=4.0.0
pyjwt>=2.8.0

# ── HTTP & Utilities ──────────────────────────────────────
httpx>=0.27.0
rich>=13.0

# ── Optional: SRT subtitle generation (requires ffmpeg) ──
# Uncomment if you need Whisper-based subtitle generation:
# openai-whisper>=20231117
